+++ src/GitLabApiClient/Models/Pipelines/Responses/TestReport.cs
@@ -6,7 +6,7 @@ namespace GitLabApiClient.Models.Pipelines.Responses
     public class TestReport
     {
         [JsonProperty("total_time")]
-        public int TotalTime { get; set; }
+        public float TotalTime { get; set; }
 
         [JsonProperty("total_count")]
         public int TotalCount { get; set; }
